[FIX] Move prompts out of trading view controller in order to remove force unwraps
